[refactoring] refactoring 1/n
