+++ TradingSystem/test.cpp
@@ -1,117 +1,15 @@
+#pragma once
 #include "gmock/gmock.h"
 #include "custom_exception.cpp"
-#include <windows.h>
+#include "driver.cpp"
+#include "auto_trading_system.cpp"
 
 using namespace testing;
-
-class Driver {
-public:
-	virtual bool login(std::string ID, std::string pass) = 0;
-	virtual int getPrice(std::string code) = 0;
-	virtual void buy(std::string code, int count, int price) = 0;
-	virtual void sell(std::string code, int price, int quantity) = 0;
-	virtual int getAvailableCash() = 0;
-	virtual void depositCash(int cash) = 0;
-	virtual int getAvailableShares(std::string code) = 0;
-};
-
-class MockDriver : public Driver {
-public:
-	MOCK_METHOD(bool, login, (std::string ID, std::string pass), (override));
-	MOCK_METHOD(int, getPrice, (std::string code), (override));
-	MOCK_METHOD(void, buy, (std::string code, int count, int price), (override));
-	MOCK_METHOD(void, sell, (std::string code, int price, int quantity), (override));
-	MOCK_METHOD(int, getAvailableCash, (), (override));
-	MOCK_METHOD(void, depositCash, (int cash), (override));
-	MOCK_METHOD(int, getAvailableShares, (std::string code), (override));
-};
-
-class StockBrockerDriverInterface {
-public:
-	std::string UNKNOWN_STOCK_CODE = "Unknown";
-	void selectStockBrocker(Driver& driver) {
-		this->driver = &driver;
-	}
-	bool login(std::string ID, std::string pass) {
-		if (driver == nullptr) {
-			throw UnknownDriverException();
-		}
-
-		return driver->login(ID, pass);
-	}
-	int getPrice(std::string code) {
-		CheckStockCode(code);
-		return driver->getPrice(code);
-	}
-	bool buy(std::string code, int count, int price) {
-		CheckStockCode(code);
-		if (driver->getAvailableCash() < price) {
-			throw InsufficientBalanceException();
-		}
-
-		driver->buy(code, count, price);
-		return true;
-	}
-	bool sell(std::string code, int price, int quantity) {
-		CheckStockCode(code);
-		if (quantity > getAvailableShares(code)) {
-			throw InsufficientSharesException();
-			return false;
-		}
-		driver->sell(code, price, quantity);
-		return true;
-	}
-	int getAvailableCash() {
-		return driver->getAvailableCash();
-	}
-	void depositCash(int cash) {
-		driver->depositCash(cash);
-	}
-	int getAvailableShares(std::string code) {
-		return driver->getAvailableShares(code);
-	}
-	bool buyNiceTiming(std::string code, int totalPrice) {
-		int price1 = getPrice(code);
-
-		Sleep(200);
-		int price2 = getPrice(code);
-
-		Sleep(200);
-		int price3 = getPrice(code);
-
-		if (price2 <= price1 || price3 <= price2) return false;
-
-		return buy(code, price3, totalPrice / price3);
-	}
-	bool sellNiceTiming(std::string code, int quantity) {
-		int price1 = getPrice(code);
-
-		Sleep(200);
-		int price2 = getPrice(code);
-
-		Sleep(200);
-		int price3 = getPrice(code);
-
-		if (price1 <= price2 || price2 <= price3) return false;
-
-		return sell(code, price3, quantity);
-	}
-private:
-	Driver* driver = nullptr;
-
-	void CheckStockCode(std::string& code)
-	{
-		if (code == UNKNOWN_STOCK_CODE) {
-			throw UnknownCodeException();
-		}
-	}
-};
-
 
 class TradingFixture :public testing::Test {
 public:
 	MockDriver mock;
-	StockBrockerDriverInterface stockerBrocker;
+	AutoTradingSystem stockerBrocker;
 	std::string id = "id1234";
 	std::string UNKNOWN = stockerBrocker.UNKNOWN_STOCK_CODE;
 	std::string password = "password56";
@@ -155,7 +53,6 @@ TEST_F(TradingFixture, TestMockGetPrice) {
 
 TEST_F(TradingFixture, TestMockUnknownStockCode1) {
 	stockerBrocker.selectStockBrocker(mock);
-	//stockerBrocker.login(id, password);
 	try {
 		stockerBrocker.getPrice(UNKNOWN);
 		FAIL();
@@ -164,7 +61,6 @@ TEST_F(TradingFixture, TestMockUnknownStockCode1) {
 }
 TEST_F(TradingFixture, TestMockUnknownStockCode2) {
 	stockerBrocker.selectStockBrocker(mock);
-	//stockerBrocker.login(id, password);
 	try {
 		stockerBrocker.buy(UNKNOWN, price, quantity);
 		FAIL();
@@ -174,7 +70,6 @@ TEST_F(TradingFixture, TestMockUnknownStockCode2) {
 
 TEST_F(TradingFixture, TestMockUnknownStockCode3) {
 	stockerBrocker.selectStockBrocker(mock);
-	//stockerBrocker.login(id, password);
 	try {
 		stockerBrocker.sell(UNKNOWN, price, quantity);
 		FAIL();
@@ -199,7 +94,6 @@ TEST_F(TradingFixture, TestMockBuyFail) {
 	EXPECT_CALL(mock, getAvailableCash())
 		.WillOnce(testing::Return(0));
 	stockerBrocker.selectStockBrocker(mock);
-	//stockerBrocker.login(id, password);
 	try {
 		stockerBrocker.buy(code, price, quantity);
 		FAIL();
@@ -213,7 +107,6 @@ TEST_F(TradingFixture, TestMockBuySuccess) {
 		.WillRepeatedly(testing::Return(cash));
 	EXPECT_CALL(mock, buy(code, price, quantity)).Times(1);
 	stockerBrocker.selectStockBrocker(mock);
-	//stockerBrocker.login(id, password);
 	stockerBrocker.buy(code, price, quantity);
 }
 
@@ -252,7 +145,6 @@ TEST_F(TradingFixture, TestBuyNiceTiming) {
 		.WillOnce(testing::Return(10000));
 	EXPECT_CALL(mock, buy(code, 10000, 100)).Times(1);
 	stockerBrocker.selectStockBrocker(mock);
-	//stockerBrocker.login(id, password);
 	stockerBrocker.buyNiceTiming(code, cash);
 }
 
@@ -266,7 +158,6 @@ TEST_F(TradingFixture, TestBuyNiceTiming2) {
 		.WillOnce(testing::Return(9000));
 	EXPECT_CALL(mock, buy(code, 10000, 100)).Times(0);
 	stockerBrocker.selectStockBrocker(mock);
-	//stockerBrocker.login(id, password);
 	stockerBrocker.buyNiceTiming(code, cash);
 }
 
@@ -279,7 +170,6 @@ TEST_F(TradingFixture, TestSellNiceTiming) {
 		.WillOnce(testing::Return(9000));
 	EXPECT_CALL(mock, sell(code, 9000, 100)).Times(1);
 	stockerBrocker.selectStockBrocker(mock);
-	//stockerBrocker.login(id, password);
 	stockerBrocker.sellNiceTiming(code, 100);
 }
 
@@ -292,6 +182,5 @@ TEST_F(TradingFixture, TestSellNiceTiming2) {
 		.WillOnce(testing::Return(10000));
 	EXPECT_CALL(mock, sell).Times(0);
 	stockerBrocker.selectStockBrocker(mock);
-	//stockerBrocker.login(id, password);
 	stockerBrocker.sellNiceTiming(code, 100);
 }
